+++ src/Stack/Options/GlobalParser.hs
@@ -3,6 +3,7 @@
 
 module Stack.Options.GlobalParser where
 
+import           Data.Array.IArray ((//))
 import           Options.Applicative
 import           Options.Applicative.Builder.Extra
 import qualified Stack.Docker                      as Docker
@@ -16,6 +17,7 @@ import           Stack.Types.Config
 import           Stack.Types.Docker
 import           Stack.Types.PrettyPrint (Styles)
 import           Stack.Types.Runner
+import           Stack.Types.StylesUpdate (StylesUpdate (..))
 
 -- | Parser for global command-line options.
 globalOptsParser :: FilePath -> GlobalOptsContext -> Maybe LogLevel -> Parser GlobalOptsMonoid
@@ -44,6 +46,17 @@ globalOptsParser currentDir kind defLogLevel =
               \that do not support color codes, the default is 'never'; color \
               \may work on terminals that support color codes" <>
          hide)) <*>
+    optionalFirst (option readStyles
+         (long "stack-colors" <>
+          metavar "STYLES" <>
+          help "Specify stack's output styles; STYLES is a colon-delimited \
+               \sequence of key=value, where 'key' is a style name and 'value' \
+               \is a semicolon-delimited list of 'ANSI' SGR (Select Graphic \
+               \Rendition) control codes (in decimal). Use 'stack ls \
+               \stack-colors --basic' to see the current sequence. In shells \
+               \where a semicolon is a command separator, enclose STYLES in \
+               \quotes." <>
+          hide)) <*>
     optionalFirst (option auto
         (long "terminal-width" <>
          metavar "INT" <>
@@ -73,7 +86,8 @@ globalOptsFromMonoid defaultTerminal defaultColorWhen defaultStyles GlobalOptsMo
     , globalCompiler = getFirst globalMonoidCompiler
     , globalTerminal = fromFirst defaultTerminal globalMonoidTerminal
     , globalColorWhen = fromFirst defaultColorWhen globalMonoidColorWhen
-    , globalStyles = defaultStyles
+    , globalStyles = defaultStyles // stylesUpdate
+                       (fromFirst (StylesUpdate []) globalMonoidStyles)
     , globalTermWidth = getFirst globalMonoidTermWidth
     , globalStackYaml = maybe SYLDefault SYLOverride $ getFirst globalMonoidStackYaml }